requirements.txt
# Core dependencies
pyyaml>=6.0
requests>=2.26.0

# REST API dependencies
flask>=2.3.0
flask-cors>=4.0.0

# Optional: For better performance
gunicorn>=21.0.0